LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/reload-intercept_no-currententrychange-expected.txt


FAIL event and promise ordering for navigation.reload() intercepted by intercept() assert_array_equals: expected property 2 to be "navigatesuccess" but got "committed fulfilled" (expected array ["navigate", "handler run", "navigatesuccess", "committed fulfilled", "promise microtask", "finished fulfilled", "transition.finished fulfilled"] got ["navigate", "handler run", "committed fulfilled", "promise microtask", "navigatesuccess", "finished fulfilled", "transition.finished fulfilled"])
